% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbrmiscpp-package.R
\docType{package}
\name{hrbrmiscpp}
\alias{hrbrmiscpp}
\alias{hrbrmiscpp-package}
\title{R C/C++ mish-mash of stuff for hrbrmstr}
\description{
R C/C++ mish-mash of stuff for hrbrmstr
}
\author{
Bob Rudis (@hrbrmstr)
}
