% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{line_count}
\alias{line_count}
\title{Count lines (fast, like "wc")}
\usage{
line_count(f)
}
\arguments{
\item{f}{full path to file}
}
\description{
Count lines (fast, like "wc")
}
\examples{
line_count(system.file("extdata", "lines.csv", package = "hrbrmiscpp"))
}
