% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdbm_hash}
\alias{sdbm_hash}
\title{SDBM Hash Function}
\usage{
sdbm_hash(x)
}
\arguments{
\item{x}{character vector of strings to hash}
}
\value{
numeric vector of the hashes of \code{x}
}
\description{
This algorithm was created for sdbm (a public-domain reimplementation of \code{ndbm})
database library. it was found to do well in scrambling bits, causing better
distribution of the keys and fewer splits. It also happens to be a good general
hashing function with good distribution.
}
\examples{
setdiff(sdbm_hash(state.name), sdbm_hash(state.name))
## numeric(0)
}
\references{
\url{http://www.cse.yorku.ca/~oz/hash.html}
}
