% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{djb2_hash}
\alias{djb2_hash}
\title{DJB2 Hash Function}
\usage{
djb2_hash(x)
}
\arguments{
\item{x}{character vector of strings to hash}
}
\value{
numeric vector of the hashes of \code{x}
}
\description{
DJB2 Hash Function
}
\examples{
setdiff(djb2_hash(state.name), djb2_hash(state.name))
## numeric(0)
}
